+++ ethylene-toml/src/test/java/com/github/steanky/ethylene/codec/toml/TomlCodecTest.java
@@ -1,0 +1,7 @@
+package com.github.steanky.ethylene.codec.toml;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TomlCodecTest {
+
+}
